# api/openapi/openapi.v1.yaml
openapi: 3.0.3
info:
  title: cixing api
  version: 1.0.0
  description: |
    Cixing ("姝よ") API V1.0.
servers:
  - url: /
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    X-Request-ID:
      description: Correlation id for tracing.
      schema:
        type: string
  parameters:
    XRequestId:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
      description: Optional client-provided request id for tracing.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
      description: Optional idempotency key.
  responses:
    BadRequest:
      description: Bad Request (validation failed)
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    Unauthorized:
      description: Unauthorized
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not Found
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal Server Error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorCode:
      type: string
      enum: [validation, unauthorized, forbidden, not_found, conflict, feature_not_ready, internal]
    ErrorResponse:
      type: object
      required: [code, reason, message, request_id]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        reason:
          type: string
          example: "common.unauthorized"
        message:
          type: string
        request_id:
          type: string
    ValidationErrorDetail:
      type: object
      required: [field, rule, reason]
      properties:
        field:
          type: string
          example: "email"
        rule:
          type: string
          example: "email"
        reason:
          type: string
          example: "invalid format"
    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            code:
              type: string
              enum: [validation]
            details:
              type: array
              items:
                $ref: "#/components/schemas/ValidationErrorDetail"

    # ---- Common Enums ----
    KeywordCategory:
      type: string
      enum: [emotion, color, shape, time, abstract]
    PromptKind:
      type: string
      enum: [intuition, structure, concept]
    ReactionType:
      type: string
      enum: [inspired, resonated]
...
    OfficialKeyword:
      type: object
      required: [id, text, category, is_active, display_order]
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
          minLength: 1
          maxLength: 50
        category:
          $ref: "#/components/schemas/KeywordCategory"
        is_active:
          type: boolean
        display_order:
          type: integer
          minimum: 1
    ReactionCounts:
      type: object
      required: [inspired, resonated]
      properties:
        inspired:
          type: integer
          minimum: 0
        resonated:
          type: integer
          minimum: 0
    TodayKeywordResponse:
      type: object
      required: [biz_date, keyword, participant_user_count]
      properties:
        biz_date:
          type: string
          format: date
        keyword:
          $ref: "#/components/schemas/OfficialKeyword"
        participant_user_count:
          type: integer
          minimum: 0
    OfficialHomeResponse:
      type: object
      required: [today, yesterday]
      properties:
        today:
          $ref: "#/components/schemas/TodayKeywordResponse"
        yesterday:
          $ref: "#/components/schemas/TodayKeywordResponse"
    Prompt:
      type: object
      required: [id, kind, content]
      properties:
        id:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/PromptKind"
        content:
          type: string
          minLength: 1
          maxLength: 200
    DrawPromptRequest:
      type: object
      required: [kind]
      properties:
        kind:
          $ref: "#/components/schemas/PromptKind"

    # ---- Images & Uploads ----
    ImageVariant:
      type: object
      required: [url, width, height]
      properties:
        url:
          type: string
          format: uri
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
...
    ImageRefSquareSmall:
      type: object
      required: [id, variants]
      properties:
        id:
          type: string
          format: uuid
        variants:
          type: object
          required: [square_small]
          properties:
            square_small:
              $ref: "#/components/schemas/ImageVariant"
    ImageRefSquareMedium:
      type: object
      required: [id, variants]
      properties:
        id:
          type: string
          format: uuid
        variants:
          type: object
          required: [square_medium]
          properties:
            square_medium:
              $ref: "#/components/schemas/ImageVariant"
    ImageRefCard4x3:
      type: object
      required: [id, variants]
      properties:
        id:
          type: string
          format: uuid
        variants:
          type: object
          required: [card_4x3]
          properties:
            card_4x3:
              $ref: "#/components/schemas/ImageVariant"
...
    ImageRefDetailLarge:
      type: object
      required: [id, variants]
      properties:
        id:
          type: string
          format: uuid
        variants:
          type: object
          required: [detail_large]
          properties:
            detail_large:
              $ref: "#/components/schemas/ImageVariant"
...
    SetCustomKeywordCoverRequest:
      type: object
      required: [image_id]
      properties:
        image_id:
          type: string
          format: uuid
    SetCustomKeywordCoverResponse:
      type: object
      required: [keyword_id, cover_source]
      properties:
        keyword_id:
          type: string
          format: uuid
        cover_source:
          $ref: "#/components/schemas/CustomKeywordCoverSource"
        cover_image:
          $ref: "#/components/schemas/ImageRefDetailLarge"
          nullable: true

    # ---- WorkUploads / WorkImages ----
    WorkImage:
      type: object
      required: [id, image, display_order, has_audio, created_at]
      properties:
        id:
          type: string
          format: uuid
        image:
          $ref: "#/components/schemas/ImageRefDetailLarge"
        display_order:
          type: integer
          minimum: 1
        title:
          type: string
          maxLength: 80
          nullable: true
        note:
          type: string
          maxLength: 500
          nullable: true
        has_audio:
          type: boolean
        audio_duration_ms:
          type: integer
          minimum: 1
          nullable: true
        audio_play_url:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
    PublicWorkUploadCard:
      type: object
      required: [id, biz_date, keyword_id, cover_image, image_count, created_at]
      properties:
        id:
          type: string
          format: uuid
        biz_date:
          type: string
          format: date
        keyword_id:
          type: string
          format: uuid
        cover_image:
          $ref: "#/components/schemas/ImageRefCard4x3"
        display_text:
          type: string
          maxLength: 500
          nullable: true
          description: UI text for list card, fallback policy decided by server (title first, else note).
        cover_has_audio:
          type: boolean
          nullable: true
        cover_audio_duration_ms:
          type: integer
          minimum: 1
          nullable: true
        image_count:
          type: integer
          minimum: 1
        reaction_counts:
          $ref: "#/components/schemas/ReactionCounts"
          nullable: true
        my_reactions:
          type: array
          items:
            $ref: "#/components/schemas/ReactionType"
          nullable: true
        created_at:
          type: string
          format: date-time
    PublicWorkUploadDetail:
      allOf:
        - $ref: "#/components/schemas/PublicWorkUploadCard"
        - type: object
          required: [images]
          properties:
            images:
              type: array
              items:
                $ref: "#/components/schemas/WorkImage"
    ListPublicUploadsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PublicWorkUploadCard"
        seed:
          type: number
          format: double
          nullable: true
...
    CustomKeywordCoverSource:
      type: string
      enum: [auto_latest, manual]
    CustomKeywordImageCard:
      type: object
      required: [id, image, display_order, created_at]
      properties:
        id:
          type: string
          format: uuid
        image:
          $ref: "#/components/schemas/ImageRefSquareMedium"
        display_order:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
    ListCustomKeywordImagesResponse:
      type: object
      required: [cover_source, items]
      properties:
        cover_source:
          $ref: "#/components/schemas/CustomKeywordCoverSource"
        cover_image:
          $ref: "#/components/schemas/ImageRefDetailLarge"
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/CustomKeywordImageCard"
    CustomKeywordImageDetail:
      type: object
      required: [id, custom_keyword_id, image, display_order, has_audio, created_at]
      properties:
        id:
          type: string
          format: uuid
        custom_keyword_id:
          type: string
          format: uuid
        image:
          $ref: "#/components/schemas/ImageRefDetailLarge"
        display_order:
          type: integer
          minimum: 1
        title:
          type: string
          maxLength: 80
          nullable: true
        note:
          type: string
          maxLength: 500
          nullable: true
        has_audio:
          type: boolean
        audio_duration_ms:
          type: integer
          minimum: 1
          nullable: true
        audio_play_url:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
...
    CustomKeywordItem:
      type: object
      required: [id, text, is_active, total_image_count, my_image_count, created_at]
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
          maxLength: 80
        target_image_count:
          type: integer
          minimum: 1
          nullable: true
        is_active:
          type: boolean
        total_image_count:
          type: integer
          minimum: 0
        my_image_count:
          type: integer
          minimum: 0
        cover_image:
          $ref: "#/components/schemas/ImageRefSquareSmall"
          nullable: true
        created_at:
          type: string
          format: date-time
    ListCustomKeywordsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CustomKeywordItem"
    CreateCustomKeywordRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 80
        target_image_count:
          type: integer
          minimum: 1
          nullable: true
    UpdateCustomKeywordRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 80
        target_image_count:
          type: integer
          minimum: 1
          nullable: true
        is_active:
          type: boolean

    # ---- Reactions ----
...
paths:
...
  /v1/official/home:
    get:
      tags: [Official]
      operationId: getOfficialHome
      security: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Optional base biz_date (defaults to today). Response includes base day as today and previous day as yesterday.
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficialHomeResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/official/keywords/{keyword_id}/prompts/draw:
    post:
      tags: [Official]
      operationId: drawOfficialPrompt
      security: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - name: keyword_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrawPromptRequest"
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---- Upload publish sessions (Object storage direct upload) ----
...
  /v1/custom-keywords:
    get:
      tags: [CustomKeywords]
      operationId: listCustomKeywords
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListCustomKeywordsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [CustomKeywords]
      operationId: createCustomKeyword
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomKeywordRequest"
      responses:
        "201":
          description: Created
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomKeywordItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/custom-keywords/{keyword_id}:
    patch:
      tags: [CustomKeywords]
      operationId: updateCustomKeyword
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: keyword_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomKeywordRequest"
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomKeywordItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [CustomKeywords]
      operationId: deleteCustomKeyword
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: keyword_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: No Content
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/custom-keywords/{keyword_id}/images:
    get:
      tags: [CustomKeywords]
      operationId: listCustomKeywordImages
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - name: keyword_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListCustomKeywordImagesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/custom/images/{image_id}:
    get:
      tags: [CustomKeywords]
      operationId: getCustomImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - name: image_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomKeywordImageDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/custom-keywords/{keyword_id}/cover:
    patch:
      tags: [CustomKeywords]
      operationId: setCustomKeywordCover
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: keyword_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCustomKeywordCoverRequest"
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetCustomKeywordCoverResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [CustomKeywords]
      operationId: clearCustomKeywordCover
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: keyword_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetCustomKeywordCoverResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---- Review ----
...
  /v1/official/dates/{biz_date}/uploads:
    get:
      tags: [Official]
      operationId: listOfficialDateUploads
      security: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - name: biz_date
          in: path
          required: true
          schema:
            type: string
            format: date
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [latest, random]
            default: latest
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: seed
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: include_reaction_counts
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Whether to include reaction_counts in response items.
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListPublicUploadsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/official/uploads/{upload_id}:
    get:
      tags: [Official]
      operationId: getOfficialUpload
      security: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - name: upload_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: include_reaction_counts
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicWorkUploadDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---- Reactions ----
...